add NEM and Interconnection limits to electric_utility_constraints.jl
adding legacy binNMIL only as needed: binNMIL was used in concert with redundant "tech" models for the "regime" above NEM. The model had to choose either techs that could NEM or not, and only the NEM techs could export into the NEM bin (and only the above NEM techs could export into the WHL bin. Now, it is simply a choice between NEM or WHL benefits. When NEM is chosen an indicator constraint enforces the net metering capacity limit.